Fix layer violations: app no longer imports from infra (#347)
Move config DTOs (ConfigSessionMode, ConfigAgentRuntime, ConfigContextConfig)
to domain::config_types. Move bus wire types (BusMessage, BusEnvelope) to
ports::bus_wire. Replace direct infra imports in app/ with domain/ports paths
or app-layer factory functions (connect_bus, new_context_store).

Remaining app→infra references are limited to composition-root modules
(app/bus.rs, app/task.rs, app/statemachine.rs) and test-only imports.

Closes #347
