+++ src/app/task.rs
@@ -1,3 +1,6 @@
-//! Backward-compatible re-export — implementation moved to `infra::task_store`.
+//! Task store application layer — re-exports task store types.
+//!
+//! The task store implementation lives in `infra::task_store`. This module
+//! re-exports it for backward compatibility.
 
 pub use crate::infra::task_store::*;
