+++ src/config.rs
@@ -1,13 +1,14 @@
-use crate::infra::dto::{
-    ConfigAgentRuntime, ConfigContextConfig, ConfigModelDef, ConfigSessionMode,
-};
+use crate::domain::config_types::{ConfigAgentRuntime, ConfigContextConfig, ConfigSessionMode};
+use crate::infra::dto::ConfigModelDef;
 use anyhow::{Context, Result};
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 use std::path::PathBuf;
 
 // Re-export path helpers (infra layer — filesystem layout concerns).
-pub use crate::infra::paths::{agent_bus_socket, log_dir, reminders_dir, state_dir};
+pub use crate::infra::paths::{
+    agent_bus_socket, ensure_dir_owned, log_dir, reminders_dir, state_dir,
+};
 
 /// A one-shot reminder that fires at a specific time and posts a message to the bus.
 #[derive(Debug, Clone, Serialize, Deserialize)]
