+++ src/infra/dto/bus.rs
@@ -1,101 +1,13 @@
 //! Bus wire-format DTOs and domain conversions.
-
-use serde::{Deserialize, Serialize};
+//!
+//! Wire types (`BusMessage`, `BusMetadata`, `BusRegister`, `BusEnvelope`) have moved
+//! to `ports::bus_wire`. This module re-exports them for backward compatibility within
+//! infra, and provides additional conversions (DomainEvent → JSON, TokenUsage parsing).
 
 use crate::domain::events::DomainEvent;
-use crate::domain::message::{Envelope, Message, Metadata, Register};
 
-// ─── Wire types ─────────────────────────────────────────────────────────────
-
-/// Wire format for messages on the Unix socket bus.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct BusMessage {
-    pub id: String,
-    pub source: String,
-    pub target: String,
-    pub payload: serde_json::Value,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub reply_to: Option<String>,
-    #[serde(default)]
-    pub metadata: BusMetadata,
-}
-
-/// Wire format for message metadata.
-#[derive(Debug, Clone, Default, Serialize, Deserialize)]
-pub struct BusMetadata {
-    #[serde(default = "default_priority")]
-    pub priority: u8,
-    #[serde(default)]
-    pub fresh: bool,
-}
-
-fn default_priority() -> u8 {
-    5
-}
-
-/// Wire format for the registration handshake.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct BusRegister {
-    pub name: String,
-    #[serde(default)]
-    pub subscriptions: Vec<String>,
-}
-
-/// Wire-level envelope: tagged union for all bus protocol messages.
-#[derive(Debug, Serialize, Deserialize)]
-#[serde(tag = "type", rename_all = "lowercase")]
-pub enum BusEnvelope {
-    Register(BusRegister),
-    Message(BusMessage),
-    List,
-}
-
-// ─── Domain ↔ Bus conversions ───────────────────────────────────────────────
-
-impl From<BusMessage> for Message {
-    fn from(dto: BusMessage) -> Self {
-        Self {
-            id: dto.id,
-            source: dto.source,
-            target: dto.target,
-            payload: dto.payload,
-            reply_to: dto.reply_to,
-            metadata: Metadata {
-                priority: dto.metadata.priority,
-                fresh: dto.metadata.fresh,
-            },
-        }
-    }
-}
-
-impl From<&Message> for BusMessage {
-    fn from(msg: &Message) -> Self {
-        Self {
-            id: msg.id.clone(),
-            source: msg.source.clone(),
-            target: msg.target.clone(),
-            payload: msg.payload.clone(),
-            reply_to: msg.reply_to.clone(),
-            metadata: BusMetadata {
-                priority: msg.metadata.priority,
-                fresh: msg.metadata.fresh,
-            },
-        }
-    }
-}
-
-impl From<BusEnvelope> for Envelope {
-    fn from(dto: BusEnvelope) -> Self {
-        match dto {
-            BusEnvelope::Register(r) => Envelope::Register(Register {
-                name: r.name,
-                subscriptions: r.subscriptions,
-            }),
-            BusEnvelope::Message(m) => Envelope::Message(m.into()),
-            BusEnvelope::List => Envelope::List,
-        }
-    }
-}
+// Re-export wire types from ports.
+pub use crate::ports::bus_wire::{BusEnvelope, BusMessage, BusMetadata, BusRegister};
 
 // ─── Domain Event → JSON ────────────────────────────────────────────────────
 
@@ -214,6 +126,7 @@ impl From<&serde_json::Value> for TokenUsage {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::domain::message::Message;
 
     #[test]
     fn test_bus_message_to_domain_roundtrip() {
@@ -243,9 +156,9 @@ mod tests {
     fn test_bus_envelope_deserialize_message() {
         let json = r#"{"type":"message","id":"1","source":"a","target":"b","payload":{}}"#;
         let env: BusEnvelope = serde_json::from_str(json).unwrap();
-        let domain: Envelope = env.into();
+        let domain: crate::domain::message::Envelope = env.into();
         match domain {
-            Envelope::Message(m) => assert_eq!(m.source, "a"),
+            crate::domain::message::Envelope::Message(m) => assert_eq!(m.source, "a"),
             _ => panic!("expected Message"),
         }
     }
@@ -254,9 +167,9 @@ mod tests {
     fn test_bus_envelope_deserialize_register() {
         let json = r#"{"type":"register","name":"cli","subscriptions":["agent:cli"]}"#;
         let env: BusEnvelope = serde_json::from_str(json).unwrap();
-        let domain: Envelope = env.into();
+        let domain: crate::domain::message::Envelope = env.into();
         match domain {
-            Envelope::Register(r) => assert_eq!(r.name, "cli"),
+            crate::domain::message::Envelope::Register(r) => assert_eq!(r.name, "cli"),
             _ => panic!("expected Register"),
         }
     }
