+++ src/infra/dto/config.rs
@@ -1,66 +1,17 @@
 //! Configuration DTOs — YAML parsing types and domain conversions.
+//!
+//! Config value types (`ConfigSessionMode`, `ConfigAgentRuntime`, `ConfigContextConfig`)
+//! have moved to `domain::config_types`. This module re-exports them for backward
+//! compatibility within infra, and provides additional config DTOs for state machine
+//! model definitions and transition parsing.
 
 use serde::{Deserialize, Serialize};
 
-use crate::domain::agent::{AgentRuntime, SessionMode};
 use crate::domain::statemachine::{ModelDef, TransitionDef};
 use crate::domain::task::TaskCriteria;
 
-// ─── SessionMode / AgentRuntime ─────────────────────────────────────────────
-
-/// Config-level session mode (serde for YAML parsing).
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, Default)]
-#[serde(rename_all = "lowercase")]
-pub enum ConfigSessionMode {
-    #[default]
-    Persistent,
-    Ephemeral,
-}
-
-/// Config-level agent runtime (serde for YAML parsing).
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, Default)]
-#[serde(rename_all = "lowercase")]
-pub enum ConfigAgentRuntime {
-    #[default]
-    Claude,
-    Acp,
-}
-
-impl From<ConfigSessionMode> for SessionMode {
-    fn from(dto: ConfigSessionMode) -> Self {
-        match dto {
-            ConfigSessionMode::Persistent => SessionMode::Persistent,
-            ConfigSessionMode::Ephemeral => SessionMode::Ephemeral,
-        }
-    }
-}
-
-impl From<&SessionMode> for ConfigSessionMode {
-    fn from(mode: &SessionMode) -> Self {
-        match mode {
-            SessionMode::Persistent => ConfigSessionMode::Persistent,
-            SessionMode::Ephemeral => ConfigSessionMode::Ephemeral,
-        }
-    }
-}
-
-impl From<ConfigAgentRuntime> for AgentRuntime {
-    fn from(dto: ConfigAgentRuntime) -> Self {
-        match dto {
-            ConfigAgentRuntime::Claude => AgentRuntime::Claude,
-            ConfigAgentRuntime::Acp => AgentRuntime::Acp,
-        }
-    }
-}
-
-impl From<&AgentRuntime> for ConfigAgentRuntime {
-    fn from(rt: &AgentRuntime) -> Self {
-        match rt {
-            AgentRuntime::Claude => ConfigAgentRuntime::Claude,
-            AgentRuntime::Acp => ConfigAgentRuntime::Acp,
-        }
-    }
-}
+// Re-export config value types from domain.
+pub use crate::domain::config_types::{ConfigAgentRuntime, ConfigContextConfig, ConfigSessionMode};
 
 // ─── ModelDef / TransitionDef ───────────────────────────────────────────────
